requirements.txt
wandb
fastai==1.0.60
tensorboardX>=1.6
ffmpeg
ffmpeg-python
yt-dlp
jupyterlab
opencv-python>=3.3.0.10
Pillow==9.1.0
torch==1.11.0
torchvision==0.12.0
ipywidgets
gradio